Chore: address review comments
- Move code out of the cmd package into the pkg directory
- Use atomic.Bool instead of writing our own
- Move files into the testdata directory
- Harden claimCPUCount to handle other potential issues
- Remove unneeded environmental overrides
